requirements.txt
flask>=3.0
pytest>=8.0
ruff>=0.4
python-dotenv>=1.0
